Modified the beam map and beam class.
- Defined new operators and methods for beam::Beam class.
- Modified the beam map of the ctcDecoder class.

The new changes are introduced to solve the problem that I had in the last commit. The last
word information used by the beam to track word formation were lost when adding the beam
to the max queue. I modified the way beams are added to the queue to preserve such information. Specifcally, I added += operator to the beam to enable combining similar beams. I
also changed the logic of update_beams_map function to reflect the new changes.