Fixed a bug that was causing the mapped scores to fall out of the specified range.
Some details:
- The code is designed such that beams are added through just one method
(update_beams_map). In addition to adding the beam to the map, the beam's score
also gets added to a score vector that is used for min-max determination. However, the previous logic did not take into
account that similar beams are combined  and hence their score accumulates.
This caused the reported max value to be less than the actual one occasionally.