Added a dictionary fst to the ctc beam class
- Added a dictionary to check if adding a char is allowed by dictionary or not.
- All ctc beam instances share the same fst dictionary, mathcer, and symbol table (static members of the class).
- The fst dictionary uses binary search to look for transitions from one state to the next.
- Beacuse of the above, the fst must be sorted.

Note 1: I used to wait till a full word is formed before checking its validity (whether
or not it is part of the dictionary). However, this results in a lot of decoding errors
if the acoustic model assigns more weight to a token that does not form a valid word.
Depending, on the implementation of the beam search this could result in the beam search
favoring continuous words if there is a penalty on invalid words. Also, the right decoding could be eliminated at early steps.
Note 2 (bug): The static member are global variables that hold objects that are scoped. After getting out of scope I get a segmentation fault.

Renamed few methods and added some comments in the rest of the files