Changes to the beam class
- Switched to using pointers to objects instead of objects
- Updated the log output of the decoding steps
- Fixed a bug that was causing segmentation error

Changes to the ctc decoder class

- Changed the constructor, it takes a path to a dictionary fst now
- Changed the qualifier of two methods from const to non-const
- Added a method to intialize the beams dictionary, symbol table from the ctc decoder constructor
- Switched to using pointers for few objects
- I am not using pruning to loop only over the tokens with highest probability

Note: The method I am using for scoring is not right. Idividual elements like beams and
the language model are working fine. However, the way I am
updating the score by multiplying probabilites and scaling
so that the score is in [0,5], is causing missing the optimal
decoding result. In this commit, I noticed that discounting
the beams by 1000 -with the existing socring method- misses up
the scores.

Added few utility functions